Fix the Makefile work when using the system install script.
Postgres 9.3 changed the PGXS infrasturcture to use the system install script,
which refuses to install the same file multiple times.

Using wildcards in the Makefile was causing one of the SQL scripts to be
double-installed and in consequence made the extension uninstallable on 9.3.

Fix this according to the proposal in
http://www.postgresql.org/message-id/5192375A.4050101@dunslane.net (the entire
thread deals with this particular problem).